+++ packages/fabric/editor-core/src/nodeviews/ui/mediaSingle/styled.ts
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import { Layout } from '@atlaskit/editor-common';
 
 function float(layout: Layout): string {
@@ -23,34 +23,38 @@ function clear(layout: Layout): string {
   }
 }
 
-function textAlign(layout: Layout): string {
-  switch (layout) {
-    case 'wide':
-    case 'full-width':
-    case 'center':
-      return 'center';
-    default:
-      return 'left';
-  }
-}
-
 export interface WrapperProps {
   layout: Layout;
+  width: number;
+  height: number;
 }
 
-// tslint:disable-next-line:variable-name
-export const Wrapper = styled.div`
-  padding-bottom: 8px;
-  display: block;
-  ${({ layout }: WrapperProps) => {
-    return `
-      float: ${float(layout)};
-      clear: ${clear(layout)};
-      text-align: ${textAlign(layout)};
-    `;
-  }};
-
-  & > * {
-    padding: 5px 10px 0 0;
+export const MediaSingleDimensionHelper = ({
+  layout,
+  width,
+  height,
+}: WrapperProps) => css`
+  margin: 0 auto;
+  float: ${float(layout)};
+  clear: ${clear(layout)};
+  max-width: ${width}px;
+  max-height: ${height}px;
+  width: 100%;
+  &:after {
+    content: '';
+    display: block;
+    padding-bottom: ${height / width * 100}%;
   }
 `;
+
+export const Wrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  margin: 24px auto;
+  ${MediaSingleDimensionHelper};
+  // Hack for selection outline
+  & .media-wrapper {
+    margin-left: -3px;
+  }
+`;
+Wrapper.displayName = 'WrapperMediaSingle';