+++ packages/fabric/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-test-helpers
 
+## 1.3.0
+
+- [minor] FS-1461 added ContextIdentifierProvider interface to editor [0aeea41](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0aeea41)
+- [minor] FS-1461 added ContextIdentifierProvider interface to editor [0aeea41](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0aeea41)
+
 ## 1.2.9
 
 ## 1.2.8